+++ src/views/Actividad.vue
@@ -16,7 +16,7 @@
           .row.justify-content-around.align-items-center            
             .col-sm.mb-3.mb-sm-0
               p.fw-bold.mb-0 Completar espacios
-            .col-auto
+            .col-lg-auto.mb-3
               a.boton.boton--b(:href="obtenerLink('/actividades/actividad.html')" target="_blank")
                 span Realizar
                 i.fas.fa-puzzle-piece